fixing homepage header
